backend/api/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
psycopg2-binary>=2.9.6,<3.0.0
python-dotenv>=1.0.0,<2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
